Move back some state from New controller to TubesSelection component
