+++ admin/app/components/target-profiles/tubes-selection.hbs
@@ -1,6 +1,6 @@
 <div
-  {{did-insert (fn @refreshAreas this.selectedFrameworks)}}
-  {{did-update (fn @refreshAreas this.selectedFrameworks) this.selectedFrameworkIds}}
+  {{did-insert this.refreshAreas}}
+  {{did-update this.refreshAreas this.selectedFrameworkIds}}
   class="tubes-selection"
 >
   <Card class="tubes-selection__card" @title="2. Sélection des sujets">
@@ -21,7 +21,7 @@
       </PixMultiSelect>
       <div class="vertical-delimiter"></div>
       <PixButtonUpload
-        @onChange={{@fillTubeSelectionFromFile}}
+        @onChange={{this.fillTubesSelectionFromFile}}
         @backgroundColor="transparent-light"
         @isBorderVisible={{true}}
         @size="small"
@@ -30,7 +30,7 @@
       >
         Importer une présélection de sujets
       </PixButtonUpload>
-      <div class="tubes-selection__inline-layout--count">{{@selectedTubesCount}}/{{@totalTubesCount}}</div>
+      <div class="tubes-selection__inline-layout--count">{{this.selectedTubesCount}}/{{this.totalTubesCount}}</div>
     </div>
   </Card>
 
@@ -38,9 +38,9 @@
     Aucun référentiel de sélectionné
   {{else}}
     <TargetProfiles::TubesSelection::Areas
-      @areas={{@areas}}
+      @areas={{this.areas}}
       @setLevelTube={{this.setLevelTube}}
-      @selectedTubeIds={{@selectedTubeIds}}
+      @selectedTubeIds={{this.selectedTubeIds}}
       @checkTube={{this.checkTube}}
       @uncheckTube={{this.uncheckTube}}
     />
