+++ admin/app/templates/authenticated/target-profiles/new.hbs
@@ -10,16 +10,9 @@
     <TargetProfiles::CreateTargetProfileForm
       @targetProfile={{this.model.targetProfile}}
       @frameworks={{this.model.frameworks}}
-      @selectedTubeIds={{this.selectedTubeIds}}
-      @areas={{this.areas}}
-      @tubeLevels={{this.tubeLevels}}
-      @refreshAreas={{this.refreshAreas}}
       @isDisabled={{this.isSaving}}
       @onSubmit={{this.createTargetProfile}}
       @onCancel={{this.goBackToTargetProfileList}}
-      @fillTubeSelectionFromFile={{this.fillTubeSelectionFromFile}}
-      @totalTubesCount={{this.totalTubesCount}}
-      @selectedTubesCount={{this.selectedTubesCount}}
     />
   </section>
 </main>
